Refactored out code to generate tests into GenerateTests.scala.
